refactor(phase-4): split confrerieext into extensions/confrerie/ package
Break the 1127-line confrerieext.py god-extension into mixins under
extensions/confrerie/ (stats, updates, editors, requests) with shared
config/exceptions in _common.py. Move the pure Notion API client to
src/integrations/notion.py so Discord-agnostic logic lives outside the
extension.
